runtime/pprof: implement cpu profiler bias test
This patch tries to lay the groundwork for CL 324129, see https://github.com/golang/go/issues/35057.

Tested on: Linux, Darwin

Still likely to be flaky on some of the bots for various reasons, but I
think those issues can be overcome. At this point I'd be interested in
initial feedback for this patch.

Change-Id: Ia75dcb3696fa71c7d6a06c7dd88da64c1f47a5bf